requirements.txt
python-dotenv
duckdb
pandas
requests